requirements/base.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.0
black==24.8.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
coverage==7.6.1
cryptography==43.0.0
daphne==4.1.2
dj-rest-auth==6.0.0
Django==5.1.1
django-allauth==64.2.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.3
# django-global-places==0.8.13
django-model-utils==5.0.0
django-split-settings==1.3.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-writable-nested==0.7.0
drf-yasg==1.21.7
factory_boy==3.3.1
Faker==28.4.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.0.0
jsonfield==3.1.0
kombu==5.4.1
msgpack==1.0.8
mypy-extensions==1.0.0
numpy==2.1.1
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pytest==8.3.3
pytest-cov==5.0.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
sentry-sdk==2.14.0
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
split-settings==1.0.0
sqlparse==0.5.1
swapper==1.4.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
wheel==0.44.0
whitenoise==6.7.0
zope.interface==7.1.0
